Move smoke-tests to testsuite/smoke-test/src/ol_smoke_tests.rs
